pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "fastapi[standard]>=0.115.8",
    "ignr>=2.2",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "flake8>=7.1.1",
    "pytest>=8.3.4",
    "pytest-cov>=4.1.0",
    "taskipy>=1.14.1",
]

[tool.pytest.ini_options]
pythonpath = "."

[tool.taskipy.tasks]
run = 'fastapi dev fast_api/app.py'
test = 'pytest --cov=fast_api -vv'
post_test = 'coverage html'
